Issue #67: Basic Orchestration
This enhancement allows the entire UpStage system to be run as one
application despite UpStage being a collection of independent
subsystems.

It is not intended to be production-ready in its own right, but it does
make running the whole system easier while maintaining isolation of
subsystems while in development.

This change also creates a UI subsystem that is WSGI compliant while
also being a strictly client-side application.